requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==4.2.13
django-cors-headers==3.5.0
django-filter==24.2
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.30.1
whitenoise
coverage
mixer
python-decouple